apps/mat_fan_engagement/requirements.txt
-e .

paho-mqtt==1.4.0
geopy==1.19.0
numpy==1.16.2
